Improve video progress tracking and database accuracy
Update YouTube player to use the official IFrame API for more reliable progress tracking, change the `watched_minutes` database column to `DECIMAL(8,2)` to preserve fractional minutes, and adjust route order to prevent conflicts.

+++ server/db/schema.sql
@@ -118,7 +118,7 @@ CREATE TABLE IF NOT EXISTS video_progress (
   student_id INTEGER REFERENCES students(id) ON DELETE CASCADE,
   video_id INTEGER REFERENCES videos(id) ON DELETE CASCADE,
   watch_count INTEGER DEFAULT 0,
-  watched_minutes INTEGER DEFAULT 0,
+  watched_minutes DECIMAL(8,2) DEFAULT 0,
   progress_percentage DECIMAL(5,2) DEFAULT 0,
   last_watched_at TIMESTAMP DEFAULT NOW(),
   UNIQUE(student_id, video_id)
